Make Reset command more robust

+++ Commands/Reset.command
@@ -1,7 +1,10 @@
 echo "Resetting…"
-rm -rf "$INSTALL_PREFIX"/Repository
+if [ -d "$INSTALL_PREFIX"/Repository ]; then
+	rm -rf "$INSTALL_PREFIX"/Repository
+fi
 git clone "$1" "$INSTALL_PREFIX"/Repository
 cd "$INSTALL_PREFIX"/Repository
 git remote add private "$2"
+git push --force private main
 
 echo "Done!"
